FriendsWave/requirements.txt
asn1crypto==0.24.0
backports.entry-points-selectable==1.1.1
configparser==4.0.2
contextlib2==0.6.0.post1
cryptography==2.1.4
distlib==0.3.3
django-environ==0.4.5
duplicity==0.7.17
enum34==1.1.6
fasteners==0.12.0
filelock==3.2.1
idna==2.6
importlib-metadata==2.1.2
importlib-resources==3.3.1
ipaddress==1.0.17
keyring==10.6.0d
keyrings.alt==3.0
lockfile==0.12.2
Markdown==3.1.1
monotonic==1.0
mysql-connector-python==2.1.6
mysql-utilities==1.6.4
paramiko==2.0.0
pathlib2==2.3.6
pexpect==4.2.1
platformdirs==2.0.2
psycopg2-binary==2.8.6
pyasn1==0.4.2
pycrypto==2.6.1
pygobject==3.26.1
pyodbc==4.0.17
pysqlite==2.7.0
python-decouple==3.5
pyxdg==0.25
scandir==1.10.0
SecretStorage==2.3.1
singledispatch==3.7.0
six==1.16.0
typing==3.10.0.0
virtualenv==20.10.0
zipp==1.2.0
